+++ src/data/rack.rs
@@ -5,19 +5,19 @@ use sqlx::PgPool;
 
 #[derive(Serialize, Debug, Deserialize)]
 pub struct Rack {
-    pub id: i64,
+    pub id: i32,
     pub name: String,
-    pub localisation_id: i64,
+    pub localisation_id: i32,
 }
 
 #[derive(Serialize, Debug)]
 pub struct InputRack {
     pub name: String,
-    pub localisation_id: i64,
+    pub localisation_id: i32,
 }
 
 impl Rack {
-    pub async fn get_all(pool: &PgPool, user_id: i64, plant_type: PlantType) -> Result<Vec<Rack>> {
+    pub async fn get_all(pool: &PgPool, user_id: i32, plant_type: PlantType) -> Result<Vec<Rack>> {
         Ok(sqlx::query_as!(
             Rack,
             "SELECT rack.id, name, localisation_id FROM rack
